+++ funciones.h
@@ -1,7 +1,7 @@
 void ingresarMatriz(int matriz[][4], int filas, int columnas);
 void imprimirMatriz(int matriz[][4], int filas, int columnas);
-int calcularDeterminante(int matriz[][4], int filas, int columnas);
-int parteX(int matriz[][4], int filas, int columnas);
-int parteY(int matriz[][4], int filas, int columnas);
-int parteZ(int matriz[][4], int filas, int columnas);
-void partesdivDeterminante(int determinanteprincipal, int parteprincipalx, int parteprincipalY, int parteprincipalZ);
+float calcularDeterminante(int matriz[][4], int filas, int columnas);
+float parteX(int matriz[][4], int filas, int columnas);
+float parteY(int matriz[][4], int filas, int columnas);
+float parteZ(int matriz[][4], int filas, int columnas);
+void partesdivDeterminante(float determinanteprincipal, float parteprincipalx, float parteprincipalY, float parteprincipalZ);
